Move File and NotebookLinter to linter module
