+++ src/databricks/labs/ucx/source_code/jobs.py
@@ -52,7 +52,8 @@ from databricks.labs.ucx.source_code.graph import (
 from databricks.labs.ucx.source_code.linters.context import LinterContext
 from databricks.labs.ucx.source_code.notebooks.cells import CellLanguage
 from databricks.labs.ucx.source_code.python.python_ast import MaybeTree, Tree, PythonSequentialLinter
-from databricks.labs.ucx.source_code.notebooks.sources import FileLinter, Notebook
+from databricks.labs.ucx.source_code.notebooks.sources import Notebook
+from databricks.labs.ucx.source_code.linters.files import FileLinter
 from databricks.labs.ucx.source_code.path_lookup import PathLookup
 from databricks.labs.ucx.source_code.used_table import UsedTablesCrawler
