+++ src/databricks/labs/ucx/source_code/notebooks/sources.py
@@ -1,49 +1,23 @@
 from __future__ import annotations
 
-import dataclasses
 import logging
-from collections.abc import Iterable
 from pathlib import Path
-from typing import cast
 
-from astroid import Module, NodeNG  # type: ignore
 
 from databricks.sdk.service.workspace import Language
 
-from databricks.labs.ucx.source_code.base import (
-    file_language,
-    safe_read_text,
-    Advice,
-    Advisory,
-    Failure,
-    Linter,
-)
 from databricks.labs.ucx.source_code.graph import (
-    Dependency,
     DependencyGraph,
     DependencyProblem,
     InheritedContext,
     SourceContainer,
 )
-from databricks.labs.ucx.source_code.files import LocalFile
-from databricks.labs.ucx.source_code.known import KnownList
-from databricks.labs.ucx.source_code.linters.context import LinterContext
-from databricks.labs.ucx.source_code.linters.imports import (
-    SysPathChange,
-    UnresolvedPath,
-)
-from databricks.labs.ucx.source_code.notebooks.magic import MagicLine
-from databricks.labs.ucx.source_code.python.python_ast import MaybeTree, NodeBase, PythonSequentialLinter, Tree
 from databricks.labs.ucx.source_code.notebooks.cells import (
     CellLanguage,
     Cell,
     CELL_SEPARATOR,
     NOTEBOOK_HEADER,
-    RunCell,
-    PythonCell,
-    RunCommand,
 )
-from databricks.labs.ucx.source_code.path_lookup import PathLookup
 
 logger = logging.getLogger(__name__)
 
@@ -118,262 +92,3 @@ class Notebook(SourceContainer):
 
     def __repr__(self):
         return f"<Notebook {self._path}>"
-
-
-class NotebookLinter:
-    """
-    Parses a Databricks notebook and then applies available linters
-    to the code cells according to the language of the cell.
-    """
-
-    def __init__(
-        self, notebook: Notebook, path_lookup: PathLookup, context: LinterContext, inherited_tree: Tree | None = None
-    ):
-        self._context: LinterContext = context
-        self._path_lookup = path_lookup
-        self._notebook: Notebook = notebook
-        # reuse Python linter across related files and notebook cells
-        # this is required in order to accumulate statements for improved inference
-        self._python_linter: PythonSequentialLinter = cast(PythonSequentialLinter, context.linter(Language.PYTHON))
-        if inherited_tree is not None:
-            self._python_linter.append_tree(inherited_tree)
-        self._python_trees: dict[PythonCell, Tree] = {}  # the original trees to be linted
-
-    def lint(self) -> Iterable[Advice]:
-        has_failure = False
-        for advice in self._load_tree_from_notebook(self._notebook, True):
-            if isinstance(advice, Failure):  # happens when a cell is unparseable
-                has_failure = True
-            yield advice
-        if has_failure:
-            return
-        for cell in self._notebook.cells:
-            if not self._context.is_supported(cell.language.language):
-                continue
-            if isinstance(cell, PythonCell):
-                tree = self._python_trees[cell]
-                advices = self._python_linter.lint_tree(tree)
-            else:
-                linter = self._linter(cell.language.language)
-                advices = linter.lint(cell.original_code)
-            for advice in advices:
-                yield dataclasses.replace(
-                    advice,
-                    start_line=advice.start_line + cell.original_offset,
-                    end_line=advice.end_line + cell.original_offset,
-                )
-
-    def _load_tree_from_notebook(self, notebook: Notebook, register_trees: bool) -> Iterable[Advice]:
-        for cell in notebook.cells:
-            if isinstance(cell, RunCell):
-                yield from self._load_tree_from_run_cell(cell)
-                continue
-            if isinstance(cell, PythonCell):
-                yield from self._load_tree_from_python_cell(cell, register_trees)
-                continue
-
-    def _load_tree_from_python_cell(self, python_cell: PythonCell, register_trees: bool) -> Iterable[Advice]:
-        maybe_tree = MaybeTree.from_source_code(python_cell.original_code)
-        if maybe_tree.failure:
-            yield maybe_tree.failure
-        tree = maybe_tree.tree
-        # a cell with only comments will not produce a tree
-        if register_trees:
-            self._python_trees[python_cell] = tree or Tree.new_module()
-        if not tree:
-            return
-        yield from self._load_children_from_tree(tree)
-
-    def _load_children_from_tree(self, tree: Tree) -> Iterable[Advice]:
-        assert isinstance(tree.node, Module)
-        # look for child notebooks (and sys.path changes that might affect their loading)
-        base_nodes: list[NodeBase] = []
-        base_nodes.extend(self._list_run_magic_lines(tree))
-        base_nodes.extend(SysPathChange.extract_from_tree(self._context.session_state, tree))
-        if len(base_nodes) == 0:
-            self._python_linter.append_tree(tree)
-            return
-        # append globals
-        globs = cast(Module, tree.node).globals
-        self._python_linter.append_globals(globs)
-        # need to execute things in intertwined sequence so concat and sort them
-        nodes = list(cast(Module, tree.node).body)
-        base_nodes = sorted(base_nodes, key=lambda node: (node.node.lineno, node.node.col_offset))
-        yield from self._load_children_with_base_nodes(nodes, base_nodes)
-        # append remaining nodes
-        self._python_linter.append_nodes(nodes)
-
-    @staticmethod
-    def _list_run_magic_lines(tree: Tree) -> Iterable[MagicLine]:
-
-        def _ignore_problem(_code: str, _message: str, _node: NodeNG) -> None:
-            return None
-
-        commands, _ = MagicLine.extract_from_tree(tree, _ignore_problem)
-        for command in commands:
-            if isinstance(command.as_magic(), RunCommand):
-                yield command
-
-    def _load_children_with_base_nodes(self, nodes: list[NodeNG], base_nodes: list[NodeBase]) -> Iterable[Advice]:
-        for base_node in base_nodes:
-            yield from self._load_children_with_base_node(nodes, base_node)
-
-    def _load_children_with_base_node(self, nodes: list[NodeNG], base_node: NodeBase) -> Iterable[Advice]:
-        while len(nodes) > 0:
-            node = nodes.pop(0)
-            self._python_linter.append_nodes([node])
-            if node.lineno < base_node.node.lineno:
-                continue
-            yield from self._load_children_from_base_node(base_node)
-
-    def _load_children_from_base_node(self, base_node: NodeBase) -> Iterable[Advice]:
-        if isinstance(base_node, SysPathChange):
-            yield from self._mutate_path_lookup(base_node)
-            return
-        if isinstance(base_node, MagicLine):
-            magic = base_node.as_magic()
-            assert isinstance(magic, RunCommand)
-            notebook = self._load_source_from_path(magic.notebook_path)
-            if notebook is None:
-                yield Advisory.from_node(
-                    code='dependency-not-found',
-                    message=f"Can't locate dependency: {magic.notebook_path}",
-                    node=base_node.node,
-                )
-                return
-            yield from self._load_tree_from_notebook(notebook, False)
-            return
-
-    def _mutate_path_lookup(self, change: SysPathChange) -> Iterable[Advice]:
-        if isinstance(change, UnresolvedPath):
-            yield Advisory.from_node(
-                code='sys-path-cannot-compute-value',
-                message=f"Can't update sys.path from {change.node.as_string()} because the expression cannot be computed",
-                node=change.node,
-            )
-            return
-        change.apply_to(self._path_lookup)
-
-    def _load_tree_from_run_cell(self, run_cell: RunCell) -> Iterable[Advice]:
-        path = run_cell.maybe_notebook_path()
-        if path is None:
-            return  # malformed run cell already reported
-        notebook = self._load_source_from_path(path)
-        if notebook is not None:
-            yield from self._load_tree_from_notebook(notebook, False)
-
-    def _load_source_from_path(self, path: Path | None) -> Notebook | None:
-        if path is None:
-            return None  # already reported during dependency building
-        resolved = self._path_lookup.resolve(path)
-        if resolved is None:
-            return None  # already reported during dependency building
-        # TODO deal with workspace notebooks
-        language = file_language(resolved)
-        # we only support Python notebooks for now
-        if language is not Language.PYTHON:
-            logger.warning(f"Unsupported notebook language: {language}")
-            return None
-        source = safe_read_text(resolved)
-        if not source:
-            return None
-        return Notebook.parse(path, source, language)
-
-    def _linter(self, language: Language) -> Linter:
-        if language is Language.PYTHON:
-            return self._python_linter
-        return self._context.linter(language)
-
-
-class FileLinter:
-    _NOT_YET_SUPPORTED_SUFFIXES = {
-        '.scala',
-        '.sh',
-        '.r',
-    }
-    _IGNORED_SUFFIXES = {
-        '.json',
-        '.md',
-        '.txt',
-        '.xml',
-        '.yml',
-        '.toml',
-        '.cfg',
-        '.bmp',
-        '.gif',
-        '.png',
-        '.tif',
-        '.tiff',
-        '.svg',
-        '.jpg',
-        '.jpeg',
-        '.pyc',
-        '.whl',
-        '.egg',
-        '.class',
-        '.iml',
-        '.gz',
-    }
-    _IGNORED_NAMES = {
-        '.ds_store',
-        '.gitignore',
-        '.coverage',
-        'license',
-        'codeowners',
-        'makefile',
-        'pkg-info',
-        'metadata',
-        'wheel',
-        'record',
-        'notice',
-        'zip-safe',
-    }
-
-    def __init__(
-        self,
-        dependency: Dependency,
-        path_lookup: PathLookup,
-        context: LinterContext,
-        *,
-        allow_list: KnownList | None = None,
-        inherited_tree: Tree | None = None,
-    ):
-        self._dependency = dependency
-        self._path_lookup = path_lookup
-        self._context = context
-        self._allow_list = allow_list or KnownList()
-        self._inherited_tree = inherited_tree
-
-    def lint(self) -> Iterable[Advice]:
-        """Lint the file."""
-        if self._dependency.path.suffix.lower() in self._IGNORED_SUFFIXES:
-            return
-        if self._dependency.path.name.lower() in self._IGNORED_NAMES:
-            return
-        if self._dependency.path.suffix.lower() in self._NOT_YET_SUPPORTED_SUFFIXES:
-            message = f"Unsupported language for suffix: {self._dependency.path.suffix}"
-            yield Failure("unsupported-language", message, -1, -1, -1, -1)
-            return
-        source_container = self._dependency.load(self._path_lookup)
-        if isinstance(source_container, Notebook):
-            yield from self._lint_notebook(source_container)
-        elif isinstance(source_container, LocalFile):
-            yield from self._lint_file(source_container)
-        else:
-            yield Failure("unsupported-file", "Unsupported file", -1, -1, -1, -1)
-
-    def _lint_file(self, local_file: LocalFile) -> Iterable[Advice]:
-        """Lint a local file."""
-        try:
-            linter = self._context.linter(local_file.language)
-            if self._inherited_tree is not None and isinstance(linter, PythonSequentialLinter):
-                linter.append_tree(self._inherited_tree)
-            yield from linter.lint(local_file.content)
-        except ValueError:
-            # TODO: Remove when implementing: https://github.com/databrickslabs/ucx/issues/3544
-            yield Failure("unsupported-language", f"Unsupported language: {local_file.language}", -1, -1, -1, -1)
-
-    def _lint_notebook(self, notebook: Notebook) -> Iterable[Advice]:
-        """Lint a notebook."""
-        notebook_linter = NotebookLinter(notebook, self._path_lookup, self._context, self._inherited_tree)
-        yield from notebook_linter.lint()
